Refactor: Miinistry Page for Admin
Removed old minitry management for admin, replaced by the coordinator view with additional functionalities
